Jalter has sounds / main program objects properties getting for fight is changed from variables to the object property itself

+++ CharactersFate/Artoria.cs
@@ -6,10 +6,10 @@
 
         // Fields
         private string _name = "Artoria";
-        private double _hpMax = 5629;
+        private int _hpMax = 5629;
         private int _atkMax = 772;
         private int _defMax = 629;
-        private double _hp = 5629;
+        private int _hp = 5629;
         private int _atk = 172;
         private int _def = 529;
         private double _spCost = 140;
@@ -27,10 +27,10 @@
         internal double ClassDmgBonus { get => _classDmgBonus; set => _classDmgBonus = value; }
         internal string? LastComment { get; set; }
         internal string Name { get => _name; }
-        internal double HpMax { get => _hpMax; set => _hpMax = value; }
+        internal int HpMax { get => _hpMax; set => _hpMax = value; }
         internal int AtkMax { get => _atkMax; set => _atkMax = value; }
         internal int DefMax { get => _defMax; set => _defMax = value; }
-        internal double Hp { get => _hp; set => _hp = value; }
+        internal int Hp { get => _hp; set => _hp = value; }
         internal int Atk { get => _atk; set => _atk = value; }
         internal int Def { get => _def; set => _def = value; }
         internal double SpCost { get => _spCost; set => _spCost = value; }
